rendering data to front end
