+++ client/package.json
@@ -32,7 +32,7 @@
     "react-icons": "^4.12.0",
     "react-redux": "^9.0.4",
     "react-router-dom": "^6.20.1",
-    "styled-components": "^6.1.1"
+    "styled-components": "^6.1.8"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.23.5",
