+++ client/src/components/GetResults.tsx
@@ -1,0 +1,122 @@
+// import React from 'react'
+// import styled from 'styled-components';
+
+
+
+
+// // const Wrapper = styled.div`
+// // display: flex;
+// // flex-wrap: wrap;
+// // justify-content: center;
+// // align-items: center;
+// // padding: 4em;
+// // background: #404337;
+// // border-radius: 10px;
+// // max-width: 100%;
+// // `;
+
+// // const Name = styled.ol`
+// // display: flex;
+// // justify-content: center;
+// // align-items: center;
+// // text-align: center;
+// // background: white;
+// // border-radius: 4px;
+// // width: auto;
+// // text-wrap: wrap;
+// // min-height: 50px;
+// // color: #75472F;
+// // box-shadow: 1px 1px 4px black;
+// // `;
+
+// // const Card = styled.ul`
+// // text-align: center;
+// // width: 15%;
+// // background: #A5A58D;
+// // font-size: 1.2em;
+// // padding: 40px; 
+// // border-radius: 10px;
+// // margin: 10px;
+// // box-shadow: 5px 5px 10px black;
+// // `;
+
+// // const Item = styled.ul`
+// // display: flex;
+// // justify-content: center;
+// // align-items: center;
+// // text-align: center;
+// // background: whitesmoke;
+// // width: auto;
+// // text-wrap: wrap;
+// // min-height: 50px;
+// // color: #7E7E63;
+// // `;
+
+
+// const ViewMoreBtn = styled.button`
+// font-size: 1em;
+// margin: 1em;
+// padding: 0.25em 1em;
+// border-radius: 3px;
+// background-color: #75472F;
+// color: white;
+// &:hover {
+//   background-color: #CB9B7C;
+//   cursor: pointer;
+//   color: white;
+// }
+// `;
+
+// const GetResults = () => {
+//   const queryResult = useAppSelector((state: any) => state.queryResult);
+
+//   return (
+//     <>
+//       <Wrapper>
+//         {queryResult.map((plant, index) => (
+//           <Card key={index}>
+//             <>
+//               <Name>{plant.common_name}</Name>
+//               {/* Add the rest of your rendering logic for each plant */}
+//             </>
+//           </Card>
+//         ))}
+//       </Wrapper>
+//     </>
+//   );
+// };
+
+
+// export default GetResults;
+// // const GetResults = () => {
+
+
+
+// //   return (
+// //     <>
+// //     <Wrapper>
+// // {singlePlant.map((plant, index) => (
+// //   <Card key={index}>
+// //     <>
+// //       <Name>{plant.common_name}</Name>
+// //       <Image src={} />
+  
+// //       {expandIndex === index && (
+// //         <>
+// //           <Item>Watering: {plant.watering}</Item>
+// //           <Item>Sunlight: {plant.sunlight}</Item>
+// //           <Item>Scientfic name: {plant.scientific_name}</Item>
+// //           <Item>Cycle: {plant.cycle}</Item>
+          
+// //         </>
+// //       )}
+// //       {/* <ViewMoreBtn onClick={() => viewMoreBtn(index)}>More info</ViewMoreBtn> */}
+// //     </>
+// //   </Card>
+// // ))}
+// // </Wrapper>
+// //     </>
+// //   )
+// // }
+
+
