--- client/src/components/viewMoreBtn.tsx
@@ -1,19 +1,0 @@
-// import React from 'react'
-// import { useState } from 'react'
-// import mockDataList1 from '../mockData/plantListData';
-
-
-// const ViewMoreBtn = () => {
-
-// const [viewMore, setViewMore ] = useState(false);
-
-
-//   return (
-    
-//     <div>
-      
-//     </div>
-//   )
-// }
-
-// export default ViewMoreBtn
