+++ client/src/index.css
@@ -1,4 +1,4 @@
-/* Have not converted these styles as I did not see buttons being in use */
+/* Have not converted these styles as I did not see buttons being in use
 
 button {
   border-radius: 8px;
@@ -23,4 +23,4 @@ button:focus-visible {
   button {
     background-color: #f9f9f9;
   }
-}
+} */
