+++ client/src/pages/GetStarted.tsx
@@ -15,6 +15,29 @@ text-align:center;
   width: 100vw;
 }
 
+
+
+// button{
+//   height: 150px;
+//   width: 300px;
+//   background-color: #FFE8D6;
+//   color: #404337;
+//   border-radius: 20px;
+//   font-size: 18px;
+// }
+
+// button:hover{
+//   background-color: #404337;
+//   color: #FFE8D6;
+// }
+
+h1{
+  align-self: center;
+  color: #404337;
+}
+`
+
+const Button = styled.button `
 button{
   height: 150px;
   width: 300px;
@@ -27,13 +50,8 @@ button{
 button:hover{
   background-color: #404337;
   color: #FFE8D6;
-}
+}`
 
-h1{
-  align-self: center;
-  color: #404337;
-}
-`
 
 const GetStarted = () => {
   console.log("render");
